[fix]修改邮件发送卡死问题，规范部分代码结构
