+++ go.mod
@@ -5,6 +5,9 @@ go 1.15
 require (
 	github.com/gogf/gf/contrib/drivers/mysql/v2 v2.1.3
 	github.com/gogf/gf/v2 v2.1.3
+	github.com/golang/mock v1.6.0 // indirect
 	github.com/jordan-wright/email v4.0.1-0.20210109023952-943e75fe5223+incompatible // indirect
 	github.com/mattn/go-runewidth v0.0.12 // indirect
+	golang.org/x/mod v0.6.0 // indirect
+	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
 )
