--- internal/consts/consts.go
@@ -1,1 +1,0 @@
-package consts
