+++ internal/consts/consts_code.go
@@ -1,0 +1,12 @@
+package consts
+
+const (
+	DataExistCode = 1001 // 数据已存在
+	DataExistMsg  = "数据已存在"
+
+	DataNoExistCode = 1002    // 帐号不存在
+	DataNoExistMsg  = "数据不存在" //
+
+	AuthorizedFailCode = 1003
+	AuthorizedFailMsg  = "无效token"
+)
