+++ internal/consts/consts_openapi.go
@@ -1,6 +1,6 @@
 package consts
 
 const (
-	OpenAPITitle       = `GoFrame Demos`
+	OpenAPITitle       = `Election Demo`
 	OpenAPIDescription = `This is a simple demos HTTP server project that is using GoFrame. Enjoy 💖 `
 )
