--- internal/consts/consts_user.go
@@ -1,6 +1,0 @@
-package consts
-
-const (
-	UserSessionKey    = "UserSessionKey"
-	AccountSessionKey = "AccountSessionKey"
-)
