+++ internal/lib/crypto/crypto.go
@@ -1,0 +1,102 @@
+package crypto
+
+import (
+	"bytes"
+	"crypto/aes"
+	"crypto/cipher"
+	"crypto/md5"
+	"encoding/base64"
+	"encoding/hex"
+	"fmt"
+
+	"github.com/gogf/gf/v2/errors/gerror"
+)
+
+var (
+	AES_KEY = "12345678abcdefgh"
+	SLAT    = "%&%&*"
+)
+
+// @brief:填充明文
+func PKCS5Padding(plaintext []byte, blockSize int) []byte {
+	padding := blockSize - len(plaintext)%blockSize
+	padtext := bytes.Repeat([]byte{byte(padding)}, padding)
+	return append(plaintext, padtext...)
+}
+
+// @brief:去除填充数据
+func PKCS5UnPadding(origData []byte) []byte {
+	length := len(origData)
+	unpadding := int(origData[length-1])
+	return origData[:(length - unpadding)]
+}
+
+// @brief:AES加密
+func AesEncrypt(origData, key []byte) ([]byte, error) {
+	block, err := aes.NewCipher(key)
+	if err != nil {
+		return nil, err
+	}
+
+	//AES分组长度为128位，所以blockSize=16，单位字节
+	blockSize := block.BlockSize()
+	origData = PKCS5Padding(origData, blockSize)
+	blockMode := cipher.NewCBCEncrypter(block, key[:blockSize]) //初始向量的长度必须等于块block的长度16字节
+	crypted := make([]byte, len(origData))
+	blockMode.CryptBlocks(crypted, origData)
+	return crypted, nil
+}
+
+// @brief:AES解密
+func AesDecrypt(crypted, key []byte) ([]byte, error) {
+	if len(crypted) == 0 {
+		return nil, gerror.New("解析错误")
+	}
+	block, err := aes.NewCipher(key)
+	if err != nil {
+		return nil, err
+	}
+
+	//AES分组长度为128位，所以blockSize=16，单位字节
+	blockSize := block.BlockSize()
+	blockMode := cipher.NewCBCDecrypter(block, key[:blockSize]) //初始向量的长度必须等于块block的长度16字节
+	origData := make([]byte, len(crypted))
+	blockMode.CryptBlocks(origData, crypted)
+	origData = PKCS5UnPadding(origData)
+	return origData, nil
+}
+
+// 加盐md5用于存储密码
+func MD5_SALT(str []byte, salt []byte) string {
+	h := md5.New()
+	h.Write(salt) // 先写盐值
+	h.Write(str)
+	return hex.EncodeToString(h.Sum(nil))
+}
+
+func main() {
+	//key的长度必须是16、24或者32字节，分别用于选择AES-128, AES-192, or AES-256
+	var aeskey = []byte("12345678abcdefgh")
+	pass := []byte("vdncloud123456")
+	xpass, err := AesEncrypt(pass, aeskey)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	pass64 := base64.StdEncoding.EncodeToString(xpass)
+	fmt.Printf("加密后:%v\n", pass64)
+
+	bytesPass, err := base64.StdEncoding.DecodeString(pass64)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	tpass, err := AesDecrypt(bytesPass, aeskey)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Printf("解密后:%s\n", tpass)
+}
