+++ internal/logic/account/account.go
@@ -1,10 +1,10 @@
-package user
+package account
 
 import (
 	"context"
 	"crypto/md5"
 	"election/internal/dao"
-	"election/internal/lib"
+	"election/internal/lib/crypto"
 	"election/internal/model"
 	"election/internal/model/do"
 	"election/internal/model/entity"
@@ -34,12 +34,12 @@ func New() *sAccount {
 func (s *sAccount) Login(ctx context.Context, in model.AccountLoginInput) (*model.AccountLoginOut, error) {
 	password := in.Password
 	bytesPass, _ := base64.StdEncoding.DecodeString(password)
-	tpass, aesErr := lib.AesDecrypt([]byte(bytesPass), []byte(lib.AES_KEY))
+	tpass, aesErr := crypto.AesDecrypt([]byte(bytesPass), []byte(crypto.AES_KEY))
 	if aesErr != nil {
 		return nil, gerror.NewCode(gcode.CodeInvalidParameter, "密码解析错误")
 	}
 	//根据解析的密码，对密码进行加盐md5。与数据库进行对比
-	passwordMd5Salt := lib.MD5_SALT(tpass, []byte(lib.SLAT))
+	passwordMd5Salt := crypto.MD5_SALT(tpass, []byte(crypto.SLAT))
 	var account *entity.Accounts
 	//查询是否有对应的账号密码
 	var err = dao.Accounts.Ctx(ctx).Where(do.Accounts{
